20240625/code/requirements.txt
python-dotenv
tenacity
ipykernel
matplotlib
plotly
scikit-learn
openai
azure-cosmos
streamlit
pymongo
tiktoken
azure-identity
